Moved entities in their own folder https://github.com/achobanov/Web-Game/issues/11
