+++ Source/Client/source/objects/entities/explosion.ts
@@ -1,8 +1,8 @@
 import ExplosionPathName from "../../assets/explosion.png";
-import Entity from "./sprites/entity";
-import EventsService from "../services/events-service";
-import AssetsService from "../services/assets-service";
-import RemoveEntityEvent from "../events/remove-entity-event";
+import Entity from "../sprites/entity";
+import EventsService from "../../services/events-service";
+import AssetsService from "../../services/assets-service";
+import RemoveEntityEvent from "../../events/remove-entity-event";
 
 export default class Explosion extends Entity {
     _events: EventsService;
